organize docs-gen tests and docs
